доработала лабу 7
